add system macro to proc, change scale component to float, added empty renderer system

+++ src/provided/world_space.rs
@@ -8,4 +8,4 @@ pub struct Position(pub f32, pub f32, pub f32);
 pub struct Rotation(pub f32, pub f32, pub f32);
 
 #[derive(Debug, Component)]
-pub struct Scale(pub u32, pub u32, pub u32);
+pub struct Scale(pub f32, pub f32, pub f32);
